MINIFICPP-364: resolve issues with test extension. This will cause issues with bootstrapping. Also resolve linkage between civet and curl extensions
