+++ cmake/BuildTests.cmake
@@ -118,3 +118,4 @@ foreach(EXTENSION ${extensions})
 endforeach()
 
 add_test(NAME TestExecuteProcess COMMAND TestExecuteProcess )
+
