+++ extensions/civetweb/processors/ListenHTTP.cpp
@@ -18,7 +18,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "processors/ListenHTTP.h"
+#include "ListenHTTP.h"
 #include <uuid/uuid.h>
 #include <CivetServer.h>
 #include <stdio.h>
