+++ extensions/http-curl/tests/C2VerifyServeResults.cpp
@@ -30,7 +30,7 @@
 #include <iostream>
 #include <sstream>
 #include "HTTPClient.h"
-#include "InvokeHTTP.h"
+#include "processors/InvokeHTTP.h"
 #include "TestBase.h"
 #include "utils/StringUtils.h"
 #include "core/Core.h"
@@ -47,34 +47,13 @@
 #include "controllers/SSLContextService.h"
 #include "TestServer.h"
 #include "c2/C2Agent.h"
-#include "c2/protocols/RESTReceiver.h"
-#include "c2/protocols/RESTSender.h"
-#include "integration/HTTPIntegrationBase.h"
+#include "protocols/RESTReceiver.h"
+#include "HTTPIntegrationBase.h"
 #include "processors/LogAttribute.h"
 
-class Responder : public CivetHandler {
+class VerifyC2Server : public HTTPIntegrationBase {
  public:
-  explicit Responder(bool isSecure)
-      : isSecure(isSecure) {
-  }
-  bool handlePost(CivetServer *server, struct mg_connection *conn) {
-    std::string resp =
-        "{\"operation\" : \"heartbeat\", \"requested_operations\" : [{ \"operationid\" : 41, \"operation\" : \"stop\", \"name\" : \"invoke\"  }, "
-        "{ \"operationid\" : 42, \"operation\" : \"stop\", \"name\" : \"FlowController\"  } ]}";
-    mg_printf(conn, "HTTP/1.1 200 OK\r\nContent-Type: "
-              "text/plain\r\nContent-Length: %lu\r\nConnection: close\r\n\r\n",
-              resp.length());
-    mg_printf(conn, "%s", resp.c_str());
-    return true;
-  }
-
- protected:
-  bool isSecure;
-};
-
-class VerifyC2Heartbeat : public HTTPIntegrationBase {
- public:
-  explicit VerifyC2Heartbeat(bool isSecure)
+  explicit VerifyC2Server(bool isSecure)
       : isSecure(isSecure) {
     char format[] = "/tmp/ssth.XXXXXX";
     dir = testController.createTempDirectory(format);
@@ -82,11 +61,11 @@ class VerifyC2Heartbeat : public HTTPIntegrationBase {
 
   void testSetup() {
     LogTestController::getInstance().setDebug<utils::HTTPClient>();
-    LogTestController::getInstance().setTrace<minifi::c2::C2Agent>();
-    LogTestController::getInstance().setDebug<LogTestController>();
-    LogTestController::getInstance().setDebug<minifi::c2::RESTSender>();
-    LogTestController::getInstance().setDebug<minifi::c2::RESTProtocol>();
+    LogTestController::getInstance().setDebug<processors::InvokeHTTP>();
     LogTestController::getInstance().setDebug<minifi::c2::RESTReceiver>();
+      LogTestController::getInstance().setDebug<minifi::c2::C2Agent>();
+    LogTestController::getInstance().setDebug<processors::LogAttribute>();
+    LogTestController::getInstance().setDebug<minifi::core::ProcessSession>();
     std::fstream file;
     ss << dir << "/" << "tstFile.ext";
     file.open(ss.str(), std::ios::out);
@@ -95,16 +74,13 @@ class VerifyC2Heartbeat : public HTTPIntegrationBase {
   }
 
   void cleanup() {
-    LogTestController::getInstance().reset();
     unlink(ss.str().c_str());
   }
 
   void runAssertions() {
-    assert(LogTestController::getInstance().contains("Received Ack from Server") == true);
+    assert(LogTestController::getInstance().contains("Import offset 0") == true);
 
-    assert(LogTestController::getInstance().contains("C2Agent] [debug] Stopping component invoke") == true);
-
-    assert(LogTestController::getInstance().contains("C2Agent] [debug] Stopping component FlowController") == true);
+    assert(LogTestController::getInstance().contains("Outputting success and response") == true);
   }
 
   void queryRootProcessGroup(std::shared_ptr<core::ProcessGroup> pg) {
@@ -117,9 +93,13 @@ class VerifyC2Heartbeat : public HTTPIntegrationBase {
     std::string url = "";
     inv->getProperty(minifi::processors::InvokeHTTP::URL.getName(), url);
 
-    configuration->set("c2.rest.url", "http://localhost:8888/api/heartbeat");
-    configuration->set("c2.agent.heartbeat.period", "1000");
-    configuration->set("c2.rest.url.ack", "http://localhost:8888/api/heartbeat");
+
+    std::string port, scheme, path;
+    parse_http_components(url, port, scheme, path);
+    configuration->set("c2.agent.heartbeat.reporter.classes", "RESTReceiver");
+    configuration->set("c2.rest.listener.port", port);
+    configuration->set("c2.agent.heartbeat.period", "10");
+    configuration->set("c2.rest.listener.heartbeat.rooturi", path);
   }
 
  protected:
@@ -131,7 +111,6 @@ class VerifyC2Heartbeat : public HTTPIntegrationBase {
 
 int main(int argc, char **argv) {
   std::string key_dir, test_file_location, url;
-  url = "http://localhost:8888/api/heartbeat";
   if (argc > 1) {
     test_file_location = argv[1];
     key_dir = argv[2];
@@ -142,13 +121,9 @@ int main(int argc, char **argv) {
     isSecure = true;
   }
 
-  VerifyC2Heartbeat harness(isSecure);
+  VerifyC2Server harness(isSecure);
 
   harness.setKeyDir(key_dir);
-
-  Responder responder(isSecure);
-
-  harness.setUrl(url, &responder);
 
   harness.run(test_file_location);
