+++ extensions/http-curl/tests/CMakeLists.txt
@@ -26,14 +26,14 @@ FOREACH(testfile ${CURL_UNIT_TESTS})
   	get_filename_component(testfilename "${testfile}" NAME_WE)
   	add_executable("${testfilename}" "${testfile}")
   	target_include_directories(${testfilename} BEFORE PRIVATE ${CURL_INCLUDE_DIRS})
-  	target_include_directories(${testfilename} BEFORE PRIVATE "${CIVET_THIRDPARTY_ROOT}/include")
+  	target_include_directories(${testfilename} BEFORE PRIVATE "${CIVET_THIRDPARTY_ROOT}")
 	target_include_directories(${testfilename} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/libminifi/test/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../http-curl/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../http-curl/client/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../http-curl/processors/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../http-curl/protocols/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../http-curl/sitetosite/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../civetweb/include/")
+	target_include_directories(${testfilename} BEFORE PRIVATE "../")
+	target_include_directories(${testfilename} BEFORE PRIVATE "../client/")
+	target_include_directories(${testfilename} BEFORE PRIVATE "../processors/")
+	target_include_directories(${testfilename} BEFORE PRIVATE "../protocols/")
+	target_include_directories(${testfilename} BEFORE PRIVATE "../sitetosite/")
+	target_include_directories(${testfilename} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/extensions/civetweb/")
 	target_include_directories(${testfilename} BEFORE PRIVATE ./include)
     createTests("${testfilename}")
     target_link_libraries(${testfilename} ${CATCH_MAIN_LIB})
@@ -51,12 +51,12 @@ FOREACH(testfile ${CURL_INTEGRATION_TESTS})
   	target_include_directories(${testfilename} BEFORE PRIVATE ${CURL_INCLUDE_DIRS})
   	target_include_directories(${testfilename} BEFORE PRIVATE "${CIVET_THIRDPARTY_ROOT}/include")
 	target_include_directories(${testfilename} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/libminifi/test/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../http-curl/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../http-curl/client/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../http-curl/processors/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../http-curl/protocols/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../http-curl/sitetosite/")
-	target_include_directories(${testfilename} BEFORE PRIVATE "../civetweb/include/")
+	target_include_directories(${testfilename} BEFORE PRIVATE "../")
+	target_include_directories(${testfilename} BEFORE PRIVATE "../client/")
+	target_include_directories(${testfilename} BEFORE PRIVATE "../processors/")
+	target_include_directories(${testfilename} BEFORE PRIVATE "../protocols/")
+	target_include_directories(${testfilename} BEFORE PRIVATE "../sitetosite/")
+	target_include_directories(${testfilename} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/extensions/civetweb/")
 	target_include_directories(${testfilename} BEFORE PRIVATE ./include)
     createTests("${testfilename}")
     if (APPLE)
@@ -79,4 +79,4 @@ add_test(NAME C2VerifyServeResults COMMAND C2VerifyServeResults "${TEST_RESOURCE
 add_test(NAME C2VerifyHeartbeatAndStop COMMAND C2VerifyHeartbeatAndStop "${TEST_RESOURCES}/C2VerifyHeartbeatAndStop.yml" "${TEST_RESOURCES}/")
 add_test(NAME SiteToSiteRestTest COMMAND SiteToSiteRestTest "${TEST_RESOURCES}/TestSite2SiteRest.yml" "${TEST_RESOURCES}/" "http://localhost:8077/nifi-api/site-to-site")
 add_test(NAME ControllerServiceIntegrationTests COMMAND ControllerServiceIntegrationTests "${TEST_RESOURCES}/TestControllerServices.yml" "${TEST_RESOURCES}/")
-add_test(NAME ThreadPoolAdjust COMMAND ThreadPoolAdjust "${TEST_RESOURCES}/TestHTTPPostChunkedEncoding.yml" "${TEST_RESOURCES}/")
+add_test(NAME ThreadPoolAdjust COMMAND ThreadPoolAdjust "${TEST_RESOURCES}/ThreadPoolAdjust.yml" "${TEST_RESOURCES}/")
