+++ extensions/http-curl/tests/HttpGetIntegrationTest.cpp
